Move test from core and fix some features
